modified:   Project/app/Models/User.php 	modified:   Project/app/Providers/AuthServiceProvider.php 	modified:   Project/database/migrations/2024_01_31_201132_create_houses_table.php 	modified:   Project/database/migrations/2024_01_31_201133_create_favorites_table.php 	modified:   Project/database/migrations/2024_01_31_201134_create_house_locations_table.php 	modified:   Project/database/migrations/2024_01_31_201135_create_interiors_table.php 	modified:   Project/database/migrations/2024_01_31_201212_create_constructions_table.php 	modified:   Project/database/migrations/2024_01_31_201219_create_properties_table.php 	modified:   Project/database/migrations/2024_03_01_193706_create_house_images_table.php 	modified:   Project/package-lock.json 	modified:   Project/package.json 	modified:   Project/resources/views/map.blade.php

+++ Project/database/migrations/2024_03_01_193706_create_house_images_table.php
@@ -13,7 +13,7 @@ return new class extends Migration
     {
         Schema::create('house_images', function (Blueprint $table) {
             $table->increments("image_id")->nullable(false);
-            $table->integer("houseID")->unsigned()->nullable(false);
+            $table->bigInteger("houseID")->unsigned()->nullable(false);
             $table->binary('image')->nullable(false);
             $table->timestamps();
             $table->foreign('houseID')->references('houseID')->on('houses');
